error - using long int instead of int for index

+++ tools/src/opf2txt.c
@@ -12,7 +12,7 @@ int main(int argc, char **argv){
     
 	FILE *fpIn = NULL,*fpOut = NULL;
 	int n, ndata, nclasses, label, i,j;
-	long int id;
+	int id;
 	float aux;
 	size_t result = 0;
 
@@ -34,7 +34,7 @@ int main(int argc, char **argv){
 	fprintf(fpOut,"\n");
 	/*gravando vetor de caracteristicas*/
 	for(i = 0; i < n; i++){
-		result = fread(&id,sizeof(long int),1,fpIn);
+		result = fread(&id,sizeof(int),1,fpIn);
 		result = fread(&label,sizeof(int),1,fpIn);
 		fprintf(fpOut,"%ld %d ",id,label);
 		for(j = 0; j < ndata; j++){
